+++ docs/source/release-notes.rst
@@ -13,6 +13,7 @@ v1.0.12 (2025-xx-xx)
 ----------------------
 
 - Add RParam pd and qd in ``DCPF1`` for easy access to load
+- Bug fix in ``RoutineBase.export_csv()`` when path is specified
 
 v1.0.11 (2025-05-23)
 ----------------------